Revert "Merge pull request #5432 from eder-matheus/grt_offset"
This reverts commit 72ee0f9c4a12f65dc1f9164ae837b804e027b76e, reversing
changes made to 255ceb9f7cac617bb157fde53f35dbe1dff372ca.

+++ src/grt/src/fastroute/include/FastRoute.h
@@ -125,7 +125,7 @@ class FastRouteCore
                      int x2,
                      int y2,
                      int layer,
-                     uint16_t reducedCap,
+                     int reducedCap,
                      bool isReduce);
   void applyVerticalAdjustments(const odb::Point& first_tile,
                                 const odb::Point& last_tile,
